requirements.txt
Django==4.2.7
djangorestframework==3.14.0
rdflib==7.0.0
SPARQLWrapper==2.0.0
python-dotenv==1.0.0
requests==2.31.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
google-generativeai==0.3.1
